Extracted an interface from SpecmaticConfig
