+++ core/src/main/kotlin/io/specmatic/core/SpecmaticConfigV1V2Common.kt
@@ -1,0 +1,1780 @@
+package io.specmatic.core
+
+import com.fasterxml.jackson.annotation.*
+import com.fasterxml.jackson.databind.DatabindException
+import com.fasterxml.jackson.databind.JsonMappingException
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize
+import com.fasterxml.jackson.databind.exc.*
+import io.specmatic.conversions.SPECMATIC_BASIC_AUTH_TOKEN
+import io.specmatic.conversions.SPECMATIC_OAUTH2_TOKEN
+import io.specmatic.core.Configuration.Companion.configFilePath
+import io.specmatic.core.SourceProvider.*
+import io.specmatic.core.azure.AzureAPI
+import io.specmatic.core.config.*
+import io.specmatic.core.config.SpecmaticConfigVersion.VERSION_1
+import io.specmatic.core.config.SpecmaticConfigVersion.VERSION_2
+import io.specmatic.core.config.v3.ConsumesDeserializer
+import io.specmatic.core.config.v3.SpecExecutionConfig
+import io.specmatic.core.git.SystemGit
+import io.specmatic.core.log.logger
+import io.specmatic.core.pattern.ContractException
+import io.specmatic.core.pattern.parsedJSONObject
+import io.specmatic.core.utilities.*
+import io.specmatic.core.utilities.Flags.Companion.EXAMPLE_DIRECTORIES
+import io.specmatic.core.utilities.Flags.Companion.EXTENSIBLE_QUERY_PARAMS
+import io.specmatic.core.utilities.Flags.Companion.EXTENSIBLE_SCHEMA
+import io.specmatic.core.utilities.Flags.Companion.MAX_TEST_COUNT
+import io.specmatic.core.utilities.Flags.Companion.MAX_TEST_REQUEST_COMBINATIONS
+import io.specmatic.core.utilities.Flags.Companion.ONLY_POSITIVE
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_BASE_URL
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_ESCAPE_SOAP_ACTION
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_GENERATIVE_TESTS
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_PRETTY_PRINT
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_STUB_DELAY
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_TEST_PARALLELISM
+import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_TEST_TIMEOUT
+import io.specmatic.core.utilities.Flags.Companion.TEST_LENIENT_MODE
+import io.specmatic.core.utilities.Flags.Companion.TEST_STRICT_MODE
+import io.specmatic.core.utilities.Flags.Companion.VALIDATE_RESPONSE_VALUE
+import io.specmatic.core.utilities.Flags.Companion.getBooleanValue
+import io.specmatic.core.utilities.Flags.Companion.getIntValue
+import io.specmatic.core.utilities.Flags.Companion.getLongValue
+import io.specmatic.core.utilities.Flags.Companion.getStringValue
+import io.specmatic.core.value.JSONObjectValue
+import io.specmatic.core.value.Value
+import io.specmatic.reporter.ctrf.model.CtrfSpecConfig
+import io.specmatic.reporter.model.SpecType
+import io.specmatic.stub.isSameBaseIgnoringHost
+import io.specmatic.test.TestResultRecord.Companion.CONTRACT_TEST_TEST_TYPE
+import java.io.File
+import java.net.URI
+import java.nio.file.Path
+import kotlin.collections.filterIsInstance
+import kotlin.collections.orEmpty
+
+private const val excludedEndpointsWarning =
+    "WARNING: excludedEndpoints is not supported in Specmatic config v2. . Refer to https://specmatic.io/documentation/configuration.html#report-configuration to see how to exclude endpoints."
+
+private const val TESTS_DIRECTORY_ENV_VAR = "SPECMATIC_TESTS_DIRECTORY"
+private const val TESTS_DIRECTORY_PROPERTY = "specmaticTestsDirectory"
+private const val CUSTOM_IMPLICIT_STUB_BASE_ENV_VAR = "SPECMATIC_CUSTOM_IMPLICIT_STUB_BASE"
+private const val CUSTOM_IMPLICIT_STUB_BASE_PROPERTY = "customImplicitStubBase"
+private const val TEST_ENDPOINTS_API = "endpointsAPI"
+private const val TEST_FILTER_ENV_VAR = "filter"
+private const val TEST_FILTER_PROPERTY = "filter"
+private const val TEST_SWAGGER_UI_BASEURL_ENV_VAR = "swaggerUIBaseURL"
+private const val TEST_SWAGGER_UI_BASEURL_PROPERTY = "swaggerUIBaseURL"
+private const val TEST_BASE_URL_ENV_VAR = "testBaseURL"
+private const val TEST_BASE_URL_PROPERTY = "testBaseURL"
+private const val TEST_HOST_ENV_VAR = "host"
+private const val TEST_HOST_PROPERTY = "host"
+private const val TEST_PORT_ENV_VAR = "port"
+private const val TEST_PORT_PROPERTY = "port"
+private const val TEST_PROTOCOL_ENV_VAR = "protocol"
+private const val TEST_PROTOCOL_PROPERTY = "protocol"
+private const val TEST_FILTER_NAME_ENV_VAR = "FILTER_NAME"
+private const val TEST_FILTER_NAME_PROPERTY = "filterName"
+private const val TEST_FILTER_NOT_NAME_ENV_VAR = "FILTER_NOT_NAME"
+private const val TEST_FILTER_NOT_NAME_PROPERTY = "filterNotName"
+private const val TEST_OVERLAY_FILE_PATH_ENV_VAR = "overlayFilePath"
+private const val TEST_OVERLAY_FILE_PATH_PROPERTY = "overlayFilePath"
+
+const val APPLICATION_NAME = "Specmatic"
+const val APPLICATION_NAME_LOWER_CASE = "specmatic"
+const val CONFIG_FILE_NAME_WITHOUT_EXT = "specmatic"
+const val DEFAULT_TIMEOUT_IN_MILLISECONDS: Long = 6000L
+const val CONTRACT_EXTENSION = "spec"
+const val YAML = "yaml"
+const val WSDL = "wsdl"
+const val YML = "yml"
+const val JSON = "json"
+val CONFIG_EXTENSIONS = listOf(YAML, YML, JSON)
+val OPENAPI_FILE_EXTENSIONS = listOf(YAML, YML, JSON)
+val CONTRACT_EXTENSIONS = listOf(CONTRACT_EXTENSION, WSDL) + OPENAPI_FILE_EXTENSIONS
+const val DATA_DIR_SUFFIX = "_data"
+const val TEST_DIR_SUFFIX = "_tests"
+const val EXAMPLES_DIR_SUFFIX = "_examples"
+const val SPECMATIC_GITHUB_ISSUES = "https://github.com/specmatic/specmatic/issues"
+const val DEFAULT_WORKING_DIRECTORY = ".$APPLICATION_NAME_LOWER_CASE"
+
+const val SPECMATIC_DISABLE_TELEMETRY = "SPECMATIC_DISABLE_TELEMETRY"
+
+const val SPECMATIC_STUB_DICTIONARY = "SPECMATIC_STUB_DICTIONARY"
+
+const val MISSING_CONFIG_FILE_MESSAGE = "Config file does not exist. (Could not find file ./specmatic.json OR ./specmatic.yaml OR ./specmatic.yml)"
+
+class WorkingDirectory(private val filePath: File) {
+    constructor(path: String = DEFAULT_WORKING_DIRECTORY): this(File(path))
+
+    val path: String
+        get() {
+            return filePath.path
+        }
+}
+
+fun invalidContractExtensionMessage(filename: String): String {
+    return "The file $filename does not seem like a contract file. Valid extensions for contract files are ${CONTRACT_EXTENSIONS.joinToString(", ")}"
+}
+
+fun String.isContractFile(): Boolean {
+    return File(this).extension in CONTRACT_EXTENSIONS
+}
+
+fun String.loadContract(): Feature {
+    if(!this.isContractFile())
+        throw ContractException(invalidContractExtensionMessage(this))
+
+    return parseContractFileToFeature(File(this))
+}
+
+data class StubConfiguration(
+    private val generative: Boolean? = null,
+    private val delayInMilliseconds: Long? = null,
+    private val dictionary: String? = null,
+    private val includeMandatoryAndRequestedKeysInResponse: Boolean? = null,
+    private val startTimeoutInMilliseconds: Long? = null,
+    private val hotReload: Switch? = null,
+    private val strictMode: Boolean? = null,
+    private val baseUrl: String? = null,
+    private val customImplicitStubBase: String? = null,
+    private val filter: String? = null,
+    private val gracefulRestartTimeoutInMilliseconds: Long? = null,
+    private val https: HttpsConfiguration? = null,
+    private val lenientMode: Boolean? = null
+) {
+    fun getGenerative(): Boolean? {
+        return generative
+    }
+
+    fun getDelayInMilliseconds(): Long? {
+        return delayInMilliseconds ?: getLongValue(SPECMATIC_STUB_DELAY)
+    }
+
+    fun getDictionary(): String? {
+        return dictionary ?: getStringValue(SPECMATIC_STUB_DICTIONARY)
+    }
+
+    fun getIncludeMandatoryAndRequestedKeysInResponse(): Boolean? {
+        return includeMandatoryAndRequestedKeysInResponse
+    }
+
+    fun getStartTimeoutInMilliseconds(): Long? {
+        return startTimeoutInMilliseconds
+    }
+
+    fun getHotReload(): Switch? {
+        return hotReload
+    }
+
+    fun getStrictMode(): Boolean? {
+        return strictMode ?: getBooleanValue(Flags.STUB_STRICT_MODE, false)
+    }
+
+    fun getFilter(): String? {
+        return filter
+    }
+
+    fun getHttpsConfiguration(): HttpsConfiguration? {
+        return https
+    }
+
+    fun getGracefulRestartTimeoutInMilliseconds(): Long? {
+        return gracefulRestartTimeoutInMilliseconds
+    }
+
+    fun getBaseUrl(): String? {
+        return baseUrl
+    }
+
+    fun getCustomImplicitStubBase(): String? {
+        return customImplicitStubBase
+    }
+}
+
+data class VirtualServiceConfiguration(
+    private val host: String? = null,
+    private val port: Int? = null,
+    private val specs: List<String>? = null,
+    private val specsDirPath: String? = null,
+    private val logsDirPath: String? = null,
+    private val logMode: VSLogMode? = null,
+    private val nonPatchableKeys: Set<String> = emptySet()
+) {
+
+    enum class VSLogMode {
+        ALL,
+        REQUEST_RESPONSE
+    }
+
+    fun getHost(): String? {
+        return host
+    }
+
+    fun getPort(): Int? {
+        return port
+    }
+
+    fun getSpecs(): List<String>? {
+        return specs
+    }
+
+    fun getSpecsDirPath(): String? {
+        return specsDirPath
+    }
+
+    fun getLogsDirPath(): String? {
+        return logsDirPath
+    }
+
+    fun getLogMode(): VSLogMode? {
+        return logMode
+    }
+
+    fun getNonPatchableKeys(): Set<String> {
+        return nonPatchableKeys
+    }
+}
+
+data class WorkflowIDOperation(
+    val extract: String? = null,
+    val use: String? = null
+)
+
+interface WorkflowDetails {
+    fun getExtractForAPI(apiDescription: String): String?
+    fun getUseForAPI(apiDescription: String): String?
+
+    companion object {
+        val default: WorkflowDetails = WorkflowConfiguration()
+    }
+}
+
+data class WorkflowConfiguration(
+    private val ids: Map<String, WorkflowIDOperation> = emptyMap()
+) : WorkflowDetails {
+    private fun getOperation(operationId: String): WorkflowIDOperation? {
+        return ids[operationId]
+    }
+
+    override fun getExtractForAPI(apiDescription: String): String? {
+        return getOperation(apiDescription)?.extract
+    }
+
+    override fun getUseForAPI(apiDescription: String): String? {
+        val operation = getOperation(apiDescription) ?: getOperation("*")
+        return operation?.use
+    }
+}
+
+interface AttributeSelectionPatternDetails {
+    fun getDefaultFields(): List<String>
+    fun getQueryParamKey(): String
+
+    companion object {
+        val default: AttributeSelectionPatternDetails = AttributeSelectionPattern()
+    }
+}
+
+data class AttributeSelectionPattern(
+    @param:JsonAlias("default_fields")
+    @field:JsonAlias("default_fields")
+    private val defaultFields: List<String>? = null,
+    @param:JsonAlias("query_param_key")
+    @field:JsonAlias("query_param_key")
+    private val queryParamKey: String? = null
+) : AttributeSelectionPatternDetails {
+    override fun getDefaultFields(): List<String> {
+        return defaultFields ?: readEnvVarOrProperty(
+            ATTRIBUTE_SELECTION_DEFAULT_FIELDS,
+            ATTRIBUTE_SELECTION_DEFAULT_FIELDS
+        ).orEmpty().split(",").filter { it.isNotBlank() }
+    }
+
+    override fun getQueryParamKey(): String {
+        return queryParamKey ?: readEnvVarOrProperty(
+            ATTRIBUTE_SELECTION_QUERY_PARAM_KEY,
+            ATTRIBUTE_SELECTION_QUERY_PARAM_KEY
+        ).orEmpty()
+    }
+}
+
+data class SpecmaticConfigV1V2Common(
+    private val sources: List<Source> = emptyList(),
+    private val auth: Auth? = null,
+    private val pipeline: Pipeline? = null,
+    private val environments: Map<String, Environment>? = null,
+    private val hooks: Map<String, String> = emptyMap(),
+    private val proxy: ProxyConfig? = null,
+    private val repository: RepositoryInfo? = null,
+    private val report: ReportConfigurationDetails? = null,
+    private val security: SecurityConfiguration? = null,
+    private val test: TestConfiguration? = null,
+    private val stub: StubConfiguration? = null,
+    private val backwardCompatibility: BackwardCompatibilityConfig? = null,
+    private val virtualService: VirtualServiceConfiguration? = null,
+    private val examples: List<String>? = null,
+    private val workflow: WorkflowConfiguration? = null,
+    private val ignoreInlineExamples: Boolean? = null,
+    private val ignoreInlineExampleWarnings: Boolean? = null,
+    private val schemaExampleDefault: Boolean? = null,
+    private val fuzzy: Boolean? = null,
+    private val extensibleQueryParams: Boolean? = null,
+    private val escapeSoapAction: Boolean? = null,
+    private val prettyPrint: Boolean? = null,
+    private val additionalExampleParamsFilePath: String? = null,
+    private val attributeSelectionPattern: AttributeSelectionPattern? = null,
+    private val allPatternsMandatory: Boolean? = null,
+    private val defaultPatternValues: Map<String, Any> = emptyMap(),
+    private val version: SpecmaticConfigVersion? = null,
+    private val disableTelemetry: Boolean? = null,
+    private val logging: LoggingConfiguration? = null,
+    private val mcp: McpConfiguration? = null,
+    private val licensePath: Path? = null,
+    private val reportDirPath: Path? = null,
+    private val globalSettings: SpecmaticGlobalSettings? = null,
+) : SpecmaticConfig {
+    companion object {
+        fun getReport(specmaticConfig: SpecmaticConfigV1V2Common): ReportConfigurationDetails? {
+            return specmaticConfig.report
+        }
+
+        @JsonIgnore
+        fun getSources(specmaticConfig: SpecmaticConfigV1V2Common): List<Source> {
+            return specmaticConfig.sources
+        }
+
+        @JsonIgnore
+        fun getRepository(specmaticConfig: SpecmaticConfigV1V2Common): RepositoryInfo? {
+            return specmaticConfig.repository
+        }
+
+        @JsonIgnore
+        fun getPipeline(specmaticConfig: SpecmaticConfigV1V2Common): Pipeline? {
+            return specmaticConfig.pipeline
+        }
+
+        @JsonIgnore
+        fun getSecurityConfiguration(specmaticConfig: SpecmaticConfigV1V2Common?): SecurityConfiguration? {
+            return specmaticConfig?.security
+        }
+
+        @JsonIgnore
+        fun getWorkflowConfiguration(specmaticConfig: SpecmaticConfigV1V2Common): WorkflowConfiguration? {
+            return specmaticConfig.workflow
+        }
+
+        @JsonIgnore
+        fun getTestConfiguration(specmaticConfig: SpecmaticConfigV1V2Common): TestConfiguration? {
+            return specmaticConfig.test
+        }
+
+        @JsonIgnore
+        fun getVirtualServiceConfiguration(specmaticConfig: SpecmaticConfigV1V2Common): VirtualServiceConfiguration {
+            return specmaticConfig.virtualService ?: VirtualServiceConfiguration()
+        }
+
+        @JsonIgnore
+        fun getAllPatternsMandatory(specmaticConfig: SpecmaticConfigV1V2Common): Boolean? {
+            return specmaticConfig.allPatternsMandatory
+        }
+
+        @JsonIgnore
+        fun getIgnoreInlineExamples(specmaticConfig: SpecmaticConfigV1V2Common): Boolean? {
+            return specmaticConfig.ignoreInlineExamples
+        }
+
+        @JsonIgnore
+        fun getAttributeSelectionPattern(specmaticConfig: SpecmaticConfigV1V2Common): AttributeSelectionPattern {
+            return specmaticConfig.attributeSelectionPattern ?: AttributeSelectionPattern()
+        }
+
+        @JsonIgnore
+        fun getStubConfiguration(specmaticConfig: SpecmaticConfigV1V2Common): StubConfiguration {
+            return specmaticConfig.stub ?: StubConfiguration()
+        }
+
+        @JsonIgnore
+        fun getTestConfigOrNull(specmaticConfig: SpecmaticConfigV1V2Common): TestConfiguration? {
+            return specmaticConfig.test
+        }
+
+        @JsonIgnore
+        fun getStubConfigOrNull(specmaticConfig: SpecmaticConfigV1V2Common): StubConfiguration? {
+            return specmaticConfig.stub
+        }
+
+        @JsonIgnore
+        fun getVirtualServiceConfigOrNull(specmaticConfig: SpecmaticConfigV1V2Common): VirtualServiceConfiguration? {
+            return specmaticConfig.virtualService
+        }
+
+        @JsonIgnore
+        fun getAttributeSelectionConfigOrNull(specmaticConfig: SpecmaticConfigV1V2Common): AttributeSelectionPattern? {
+            return specmaticConfig.attributeSelectionPattern
+        }
+
+        fun getEnvironments(specmaticConfig: SpecmaticConfigV1V2Common): Map<String, Environment>? {
+            return specmaticConfig.environments
+        }
+
+        @JsonIgnore
+        fun getHooks(specmaticConfig: SpecmaticConfigV1V2Common): Map<String, String> {
+            return specmaticConfig.hooks
+        }
+
+        @JsonIgnore
+        fun getProxyConfig(specmaticConfig: SpecmaticConfigV1V2Common): ProxyConfig? {
+            return specmaticConfig.proxy
+        }
+    }
+
+    @JsonIgnore
+    override fun getLogConfigurationOrDefault(): LoggingConfiguration {
+        return this.logging ?: LoggingConfiguration.default()
+    }
+
+    @JsonIgnore
+    override fun isTelemetryDisabled(): Boolean {
+        val disableTelemetryFromEnvVarOrSystemProp = readEnvVarOrProperty(
+            SPECMATIC_DISABLE_TELEMETRY, SPECMATIC_DISABLE_TELEMETRY
+        )
+
+        return disableTelemetryFromEnvVarOrSystemProp?.toBoolean()
+            ?: (disableTelemetry == true)
+    }
+
+    @JsonIgnore
+    override fun testConfigFor(specPath: String, specType: String): Map<String, Any> {
+        return sources.flatMap { it.test.orEmpty() }.configWith(specPath, specType)
+    }
+
+    @JsonIgnore
+    override fun testConfigFor(file: File, specType: String): Map<String, Any> {
+        val resolvedTestConfig = this.sources.flatMap { it.getCanonicalTestConfigs() }
+        return resolvedTestConfig.filterIsInstance<SpecExecutionConfig.ConfigValue>().firstOrNull { config ->
+            if (config.specType != specType) return@firstOrNull false
+            config.specs.any { specPath -> File(specPath).sameAs(file) }
+        }?.config.orEmpty()
+    }
+
+    @JsonIgnore
+    override fun stubConfigFor(file: File, specType: String): Map<String, Any> {
+        val resolvedTestConfig = this.sources.flatMap { it.getCanonicalStubConfigs() }
+        return resolvedTestConfig.filterIsInstance<SpecExecutionConfig.ConfigValue>().firstOrNull { config ->
+            if (config.specType != specType) return@firstOrNull false
+            config.specs.any { specPath -> File(specPath).sameAs(file) }
+        }?.config.orEmpty()
+    }
+
+    @JsonIgnore
+    override fun stubConfigFor(specPath: String, specType: String): Map<String, Any> {
+        return sources.flatMap { it.stub.orEmpty() }.configWith(specPath, specType)
+    }
+
+    @JsonIgnore
+    private fun List<SpecExecutionConfig>.configWith(specPath: String, specType: String): Map<String, Any> {
+        return this.filterIsInstance<SpecExecutionConfig.ConfigValue>().firstOrNull {
+            it.contains(specPath, specType)
+        }?.config.orEmpty()
+    }
+
+    @JsonIgnore
+    override fun getCtrfSpecConfig(absoluteSpecPath: String, testType: String, protocol: String, specType: String): CtrfSpecConfig {
+        val source = when (testType) {
+            CONTRACT_TEST_TEST_TYPE -> testSourceFromConfig(absoluteSpecPath)
+            else -> stubSourceFromConfig(absoluteSpecPath)
+        } ?: Source()
+
+        val specPathFromConfig = when(testType) {
+            CONTRACT_TEST_TEST_TYPE -> testSpecPathFromConfigFor(absoluteSpecPath)
+            else -> stubSpecPathFromConfigFor(absoluteSpecPath)
+        }
+
+        return CtrfSpecConfig(
+            protocol = protocol,
+            specType = specType,
+            specification = specPathFromConfig.orEmpty(),
+            sourceProvider = source.provider.name,
+            repository = source.repository.orEmpty(),
+            branch = source.branch ?: "main",
+        )
+    }
+
+    @JsonIgnore
+    override fun getHotReload(): Switch? {
+        return getStubConfiguration(this).getHotReload()
+    }
+
+    @JsonIgnore
+    override fun dropExcludedEndpointsAfterVersion1(latestVersion: SpecmaticConfigVersion): SpecmaticConfigV1V2Common {
+        if (latestVersion == VERSION_1)
+            return this
+
+        logger.log("\n$excludedEndpointsWarning\n")
+
+        return this.copy(
+            report = report?.clearPresenceOfExcludedEndpoints()
+        )
+    }
+
+    @JsonIgnore
+    override fun getReport(): ReportConfiguration? {
+        return report
+    }
+
+    @JsonIgnore
+    override fun getWorkflowDetails(): WorkflowDetails? {
+        return workflow
+    }
+
+    @JsonIgnore
+    override fun getAttributeSelectionPattern(): AttributeSelectionPatternDetails {
+        return getAttributeSelectionPattern(this)
+    }
+
+    @JsonIgnore
+    override fun getGlobalSettingsOrDefault(): SpecmaticGlobalSettings {
+        return this.globalSettings ?: SpecmaticGlobalSettings()
+    }
+
+    @JsonIgnore
+    override fun stubBaseUrls(defaultBaseUrl: String): List<String> =
+        sources
+            .flatMap { source ->
+                source.stub.orEmpty().map { consumes ->
+                    baseUrlFrom(consumes, defaultBaseUrl)
+                }
+            }.distinct()
+
+    @JsonIgnore
+    override fun stubToBaseUrlList(defaultBaseUrl: String): List<Pair<String, String>> {
+        return sources.flatMap { source ->
+            source.stub.orEmpty().flatMap { consumes ->
+                when (consumes) {
+                    is SpecExecutionConfig.StringValue -> listOf(consumes.value to defaultBaseUrl)
+                    is SpecExecutionConfig.ObjectValue -> consumes.specs.map { it to consumes.toBaseUrl(defaultBaseUrl) }
+                    is SpecExecutionConfig.ConfigValue -> consumes.specs.map { it to defaultBaseUrl }
+                }
+            }
+        }
+    }
+
+    @JsonIgnore
+    override fun stubBaseUrlPathAssociatedTo(url: String, defaultBaseUrl: String): String {
+        val parsedUrl = URI(url)
+        return stubBaseUrls(defaultBaseUrl).map(::URI).firstOrNull { stubBaseUrl ->
+            isSameBaseIgnoringHost(parsedUrl, stubBaseUrl)
+        }?.path.orEmpty()
+    }
+
+    @JsonIgnore
+    override fun getStubStartTimeoutInMilliseconds(): Long {
+        return getStubConfiguration(this).getStartTimeoutInMilliseconds() ?: 20_000L
+    }
+
+    @JsonIgnore
+    private fun baseUrlFrom(
+        consumes: SpecExecutionConfig,
+        defaultBaseUrl: String
+    ): String = if (consumes is SpecExecutionConfig.ObjectValue) consumes.toBaseUrl(defaultBaseUrl) else defaultBaseUrl
+
+    override fun logDependencyProjects(azure: AzureAPI) {
+        logger.log("Dependency projects")
+        logger.log("-------------------")
+
+        sources.forEach { source ->
+            logger.log("In central repo ${source.repository}")
+
+            source.specsUsedAsTest().forEach { relativeContractPath ->
+                logger.log("  Consumers of $relativeContractPath")
+                val consumers = azure.referencesToContract(relativeContractPath)
+
+                if (consumers.isEmpty()) {
+                    logger.log("    ** no consumers found **")
+                } else {
+                    consumers.forEach {
+                        logger.log("  - ${it.description}")
+                    }
+                }
+
+                logger.newLine()
+            }
+        }
+    }
+
+    @JsonIgnore
+    override fun loadSources(useCurrentBranchForCentralRepo: Boolean): List<ContractSource> {
+        return sources.map { source ->
+            val defaultBaseUrl = getDefaultBaseUrl()
+            val stubExamplesMap = source.specToStubExamplesMap()
+            val stubPaths = source.specToStubBaseUrlMap(defaultBaseUrl).entries.map { ContractSourceEntry(it.key, it.value, exampleDirPaths = stubExamplesMap[it.key]) }
+
+            val testBaseUrlMap = source.specToTestBaseUrlMap(defaultBaseUrl)
+            val testGenerativeMap = source.specToTestGenerativeMap()
+            val testExamplesMap = source.specToTestExamplesMap()
+            val testPaths = testBaseUrlMap.entries.map { ContractSourceEntry(it.key, it.value, testGenerativeMap[it.key], exampleDirPaths = testExamplesMap[it.key]) }
+
+            val sourceMatchBranch = source.matchBranch ?: false
+            val effectiveUseCurrentBranch = useCurrentBranchForCentralRepo || sourceMatchBranch
+            val effectiveBranch = getEffectiveBranchForSource(source.branch, effectiveUseCurrentBranch)
+
+
+            when (source.provider) {
+                git -> when (source.repository) {
+                    null -> GitMonoRepo(testPaths, stubPaths, source.provider.toString())
+                    else -> GitRepo(
+                        source.repository,
+                        effectiveBranch,
+                        testPaths,
+                        stubPaths,
+                        source.provider.toString(),
+                        effectiveUseCurrentBranch,
+                        specmaticConfig = this
+                    )
+                }
+
+                filesystem -> LocalFileSystemSource(source.directory ?: ".", testPaths, stubPaths)
+
+                web -> WebSource(testPaths, stubPaths)
+            }
+        }
+    }
+
+    private fun getEffectiveBranchForSource(
+        configuredBranch: String?,
+        useCurrentBranchForCentralRepo: Boolean,
+    ): String? {
+        if (!useCurrentBranchForCentralRepo) {
+            return configuredBranch
+        }
+
+        return try {
+            val git = SystemGit()
+            val currentBranch = git.getCurrentBranchForMatchBranch()
+            logger.debug("Current branch: $currentBranch")
+
+            logger.log("Central repo branch: '$currentBranch'")
+            currentBranch
+        } catch (e: Throwable) {
+            val fallbackBranchToLog = configuredBranch?.let { "configured branch: $configuredBranch" } ?: "default"
+
+            logger.log("Could not determine current branch for --match-branch flag: ${e.message}")
+            logger.log("Falling back to $fallbackBranchToLog")
+
+            configuredBranch
+        }
+    }
+
+    @JsonIgnore
+    override fun attributeSelectionQueryParamKey(): String {
+        return getAttributeSelectionPattern().getQueryParamKey()
+    }
+
+    @JsonIgnore
+    override fun isExtensibleSchemaEnabled(): Boolean {
+        return test?.allowExtensibleSchema ?: getBooleanValue(EXTENSIBLE_SCHEMA)
+    }
+
+    @JsonIgnore
+    override fun isResiliencyTestingEnabled(): Boolean {
+        return (getResiliencyTestsEnabled() != ResiliencyTestSuite.none)
+    }
+
+    @JsonIgnore
+    override fun isOnlyPositiveTestingEnabled(): Boolean {
+        return (getResiliencyTestsEnabled() == ResiliencyTestSuite.positiveOnly)
+    }
+
+    @JsonIgnore
+    override fun isResponseValueValidationEnabled(): Boolean {
+        return test?.validateResponseValues ?: getBooleanValue(VALIDATE_RESPONSE_VALUE)
+    }
+
+    @JsonIgnore
+    override fun parsedDefaultPatternValues(): Map<String, Value> {
+        return parsedJSONObject(ObjectMapper().writeValueAsString(defaultPatternValues)).jsonObject
+    }
+
+    @JsonIgnore
+    override fun getResiliencyTestsEnabled(): ResiliencyTestSuite {
+        return (test?.resiliencyTests ?: ResiliencyTestsConfig.fromSystemProperties()).enable ?: ResiliencyTestSuite.none
+    }
+
+    @JsonIgnore
+    override fun getTestTimeoutInMilliseconds(): Long? {
+        return test?.timeoutInMilliseconds ?: getLongValue(SPECMATIC_TEST_TIMEOUT)
+    }
+
+    @JsonIgnore
+    override fun getMaxTestRequestCombinations(): Int? {
+        val configValue = if (getVersion() == VERSION_2) test?.maxTestRequestCombinations else null
+        return configValue ?: getIntValue(MAX_TEST_REQUEST_COMBINATIONS)
+    }
+
+    @JsonIgnore
+    override fun getTestStrictMode(): Boolean? {
+        return test?.strictMode ?: getStringValue(TEST_STRICT_MODE)?.toBoolean()
+    }
+
+    @JsonIgnore
+    override fun getTestLenientMode(): Boolean? {
+        return test?.lenientMode ?: getStringValue(TEST_LENIENT_MODE)?.toBoolean()
+    }
+
+    @JsonIgnore
+    override fun getTestParallelism(): String? {
+        return test?.parallelism ?: getStringValue(SPECMATIC_TEST_PARALLELISM)
+    }
+
+    @JsonIgnore
+    override fun getTestsDirectory(): String? {
+        return test?.testsDirectory
+            ?: readEnvVarOrProperty(TESTS_DIRECTORY_ENV_VAR, TESTS_DIRECTORY_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getMaxTestCount(): Int? {
+        return test?.maxTestCount ?: getIntValue(MAX_TEST_COUNT)
+    }
+
+    @JsonIgnore
+    override fun getTestFilter(): String? {
+        return getTestConfiguration(this)?.filter
+            ?: readEnvVarOrProperty(TEST_FILTER_ENV_VAR, TEST_FILTER_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestFilterName(): String? {
+        val configValue = if (getVersion() == VERSION_2) test?.filterName else null
+        return configValue ?: readEnvVarOrProperty(TEST_FILTER_NAME_ENV_VAR, TEST_FILTER_NAME_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestFilterNotName(): String? {
+        val configValue = if (getVersion() == VERSION_2) test?.filterNotName else null
+        return configValue ?: readEnvVarOrProperty(TEST_FILTER_NOT_NAME_ENV_VAR, TEST_FILTER_NOT_NAME_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestOverlayFilePath(): String? {
+        val configValue = if (getVersion() == VERSION_2) test?.overlayFilePath else null
+        return configValue ?: readEnvVarOrProperty(TEST_OVERLAY_FILE_PATH_ENV_VAR, TEST_OVERLAY_FILE_PATH_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestBaseUrl(): String? {
+        val baseUrl = getExplicitTestBaseUrl()
+        if (baseUrl != null) return baseUrl
+
+        val rawHost = readEnvVarOrProperty(TEST_HOST_ENV_VAR, TEST_HOST_PROPERTY)
+        val port = readEnvVarOrProperty(TEST_PORT_ENV_VAR, TEST_PORT_PROPERTY)
+        if (rawHost.isNullOrBlank() || port.isNullOrBlank()) return null
+
+        val host = if (rawHost.startsWith("http")) {
+            URI(rawHost).host ?: return null
+        } else {
+            rawHost
+        }
+
+        val protocol = readEnvVarOrProperty(TEST_PROTOCOL_ENV_VAR, TEST_PROTOCOL_PROPERTY) ?: "http"
+        val constructedBaseUrl = "$protocol://$host:$port"
+        return if (validateTestOrStubUri(constructedBaseUrl) == URIValidationResult.Success) {
+            constructedBaseUrl
+        } else {
+            null
+        }
+    }
+
+    @JsonIgnore
+    override fun getCoverageReportBaseUrl(): String? {
+        val baseUrl = getExplicitTestBaseUrl()
+        if (baseUrl != null) return baseUrl
+
+        val host = readEnvVarOrProperty(TEST_HOST_ENV_VAR, TEST_HOST_PROPERTY).orEmpty()
+        val port = readEnvVarOrProperty(TEST_PORT_ENV_VAR, TEST_PORT_PROPERTY).orEmpty()
+        return if (host.isNotBlank() && port.isNotBlank()) "$host:$port" else null
+    }
+
+    @JsonIgnore
+    private fun getExplicitTestBaseUrl(): String? {
+        val configValue = if (getVersion() == VERSION_2) test?.baseUrl else null
+        return configValue ?: readEnvVarOrProperty(TEST_BASE_URL_ENV_VAR, TEST_BASE_URL_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestSwaggerUrl(): String? {
+        return getTestConfiguration(this)?.swaggerUrl
+    }
+
+    @JsonIgnore
+    override fun getTestSwaggerUIBaseUrl(): String? {
+        val configValue = if (getVersion() == VERSION_2) test?.swaggerUIBaseURL else null
+        return configValue ?: readEnvVarOrProperty(TEST_SWAGGER_UI_BASEURL_ENV_VAR, TEST_SWAGGER_UI_BASEURL_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getTestJunitReportDir(): String? {
+        return if (getVersion() == VERSION_2) test?.junitReportDir else null
+    }
+
+    @JsonIgnore
+    override fun getActuatorUrl(): String? {
+        return getTestConfiguration(this)?.actuatorUrl ?: getStringValue(TEST_ENDPOINTS_API)
+    }
+
+    @JsonIgnore
+    override fun copyResiliencyTestsConfig(onlyPositive: Boolean): SpecmaticConfigV1V2Common {
+        val testConfig = test ?: TestConfiguration()
+        return this.copy(
+            test = testConfig.copy(
+                resiliencyTests = (testConfig.resiliencyTests ?: ResiliencyTestsConfig.fromSystemProperties()).copy(
+                    enable = if (onlyPositive) ResiliencyTestSuite.positiveOnly else ResiliencyTestSuite.all
+                )
+            )
+        )
+    }
+
+    @JsonIgnore
+    override fun getStubIncludeMandatoryAndRequestedKeysInResponse(): Boolean {
+        return getStubConfiguration(this).getIncludeMandatoryAndRequestedKeysInResponse() ?: true
+    }
+
+    @JsonIgnore
+    override fun getStubGenerative(): Boolean {
+        return getStubConfiguration(this).getGenerative() ?: false
+    }
+
+    @JsonIgnore
+    override fun getStubDelayInMilliseconds(): Long? {
+        return getStubConfiguration(this).getDelayInMilliseconds()
+    }
+
+    @JsonIgnore
+    override fun getStubDictionary(): String? {
+        return getStubConfiguration(this).getDictionary()
+    }
+
+    @JsonIgnore
+    override fun getStubStrictMode(): Boolean? {
+        return getStubConfiguration(this).getStrictMode()
+    }
+
+    @JsonIgnore
+    override fun getStubFilter(): String? {
+        return getStubConfiguration(this).getFilter()
+    }
+
+    @JsonIgnore
+    override fun getStubHttpsConfiguration(): HttpsConfiguration? {
+        return getStubConfiguration(this).getHttpsConfiguration()
+    }
+
+    @JsonIgnore
+    override fun getStubGracefulRestartTimeoutInMilliseconds(): Long? {
+        return getStubConfiguration(this).getGracefulRestartTimeoutInMilliseconds()
+    }
+
+    @JsonIgnore
+    override fun getDefaultBaseUrl(): String {
+        return getStubConfiguration(this).getBaseUrl()
+            ?: getStringValue(SPECMATIC_BASE_URL)
+            ?: Configuration.DEFAULT_BASE_URL
+    }
+
+    @JsonIgnore
+    override fun getCustomImplicitStubBase(): String? {
+        return getStubConfiguration(this).getCustomImplicitStubBase()
+            ?: readEnvVarOrProperty(CUSTOM_IMPLICIT_STUB_BASE_ENV_VAR, CUSTOM_IMPLICIT_STUB_BASE_PROPERTY)
+    }
+
+    @JsonIgnore
+    override fun getIgnoreInlineExamples(): Boolean {
+        return ignoreInlineExamples ?: getBooleanValue(Flags.IGNORE_INLINE_EXAMPLES)
+    }
+
+    @JsonIgnore
+    override fun getIgnoreInlineExampleWarnings(): Boolean {
+        return ignoreInlineExampleWarnings ?: getBooleanValue(Flags.IGNORE_INLINE_EXAMPLE_WARNINGS)
+    }
+
+    @JsonIgnore
+    override fun getAllPatternsMandatory(): Boolean {
+        return allPatternsMandatory ?: getBooleanValue(Flags.ALL_PATTERNS_MANDATORY)
+    }
+
+    @JsonIgnore
+    override fun getSchemaExampleDefault(): Boolean {
+        val configValue = if (getVersion() == VERSION_2) schemaExampleDefault else null
+        return configValue ?: getBooleanValue(Flags.SCHEMA_EXAMPLE_DEFAULT)
+    }
+
+    @JsonIgnore
+    override fun getFuzzyMatchingEnabled(): Boolean {
+        val configValue = if (getVersion() == VERSION_2) fuzzy else null
+        return configValue ?: getBooleanValue(Flags.SPECMATIC_FUZZY)
+    }
+
+    @JsonIgnore
+    override fun getExtensibleQueryParams(): Boolean {
+        val configValue = if (getVersion() == VERSION_2) extensibleQueryParams else null
+        return configValue ?: getBooleanValue(EXTENSIBLE_QUERY_PARAMS)
+    }
+
+    @JsonIgnore
+    override fun getEscapeSoapAction(): Boolean {
+        val configValue = if (getVersion() == VERSION_2) escapeSoapAction else null
+        return configValue ?: getBooleanValue(SPECMATIC_ESCAPE_SOAP_ACTION)
+    }
+
+    @JsonIgnore
+    override fun getPrettyPrint(): Boolean {
+        val configValue = if (getVersion() == VERSION_2) prettyPrint else null
+        return configValue ?: getBooleanValue(SPECMATIC_PRETTY_PRINT, true)
+    }
+
+    @JsonIgnore
+    override fun getAdditionalExampleParamsFilePath(): String? {
+        return additionalExampleParamsFilePath ?: getStringValue(Flags.ADDITIONAL_EXAMPLE_PARAMS_FILE)
+    }
+
+    @JsonIgnore
+    override fun getHooks(): Map<String, String> {
+        return hooks
+    }
+
+    @JsonIgnore
+    override fun getProxyConfig(): ProxyConfig? {
+        return proxy
+    }
+
+    @JsonIgnore
+    override fun getDefaultPatternValues(): Map<String, Any> {
+        return defaultPatternValues
+    }
+
+    override fun getVersion(): SpecmaticConfigVersion {
+        return this.version ?: VERSION_1
+    }
+
+    @JsonIgnore
+    override fun getMatchBranchEnabled(): Boolean {
+        return sources.any { it.matchBranch == true } || getBooleanValue(Flags.MATCH_BRANCH)
+    }
+
+    @JsonIgnore
+    override fun mapSources(transform: (Source) -> Source): SpecmaticConfigV1V2Common {
+        val transformedSources = this.sources.map(transform)
+        return this.copy(sources = transformedSources)
+    }
+
+    @JsonIgnore
+    override fun getAuth(): Auth? {
+        return auth
+    }
+
+    @JsonIgnore
+    override fun getAuthBearerFile(): String? {
+        return auth?.bearerFile
+    }
+
+    @JsonIgnore
+    override fun getAuthBearerEnvironmentVariable(): String? {
+        return auth?.bearerEnvironmentVariable
+    }
+
+    @JsonIgnore
+    override fun getAuthPersonalAccessToken(): String? {
+        val tokenFromConfig = auth?.personalAccessToken?.takeIf { it.isNotBlank() }
+        if (tokenFromConfig != null) return tokenFromConfig
+
+        val tokenFromEnv = System.getenv("PERSONAL_ACCESS_TOKEN")?.takeIf { it.isNotBlank() }
+        if (tokenFromEnv != null) {
+            logger.log("Using personal access token from environment variable")
+            return tokenFromEnv
+        }
+
+        val tokenFromProperty = System.getProperty("personalAccessToken")?.takeIf { it.isNotBlank() }
+        if (tokenFromProperty != null) {
+            logger.log("Using personal access token from system property")
+            return tokenFromProperty
+        }
+
+        return getPersonalAccessTokenFromHomeDirectoryConfig()
+    }
+
+    private fun getPersonalAccessTokenFromHomeDirectoryConfig(): String? {
+        val homeDir = File(System.getProperty("user.home"))
+        val configFile = homeDir.resolve("specmatic-azure.json")
+
+        if (!configFile.exists()) return null
+
+        val config = parsedJSONObject(configFile.readText())
+        val tokenKeys = listOf("azure-access-token", "personal-access-token")
+
+        val tokenFromHomeConfig = tokenKeys.firstNotNullOfOrNull { tokenKey ->
+            when {
+                config.jsonObject.containsKey(tokenKey) ->
+                    config.getString(tokenKey).takeIf { it.isNotBlank() }
+                else -> null
+            }
+        }
+
+        return tokenFromHomeConfig?.also {
+            logger.log("Using personal access token from home directory config")
+        }
+    }
+
+    @JsonIgnore
+    override fun getExamples(): List<String> {
+        return examples ?: getStringValue(EXAMPLE_DIRECTORIES)?.split(",") ?: emptyList()
+    }
+
+    @JsonIgnore
+    override fun getRepositoryProvider(): String? {
+        return repository?.getProvider()
+    }
+
+    @JsonIgnore
+    override fun getRepositoryCollectionName(): String? {
+        return repository?.getCollectionName()
+    }
+
+    @JsonIgnore
+    override fun getBackwardCompatibilityConfig(): BackwardCompatibilityConfig? {
+        return this.backwardCompatibility
+    }
+
+    @JsonIgnore
+    override fun getMcpConfiguration(): McpConfiguration? {
+        return this.mcp
+    }
+
+    @JsonIgnore
+    override fun getPipelineProvider(): PipelineProvider? {
+        return pipeline?.getProvider()
+    }
+
+    @JsonIgnore
+    override fun getPipelineDefinitionId(): Int? {
+        return pipeline?.getDefinitionId()
+    }
+
+    @JsonIgnore
+    override fun getPipelineOrganization(): String? {
+        return pipeline?.getOrganization()
+    }
+
+    @JsonIgnore
+    override fun getPipelineProject(): String? {
+        return pipeline?.getProject()
+    }
+
+    @JsonIgnore
+    override fun getOpenAPISecurityConfigurationScheme(scheme: String): SecuritySchemeConfiguration? {
+        return security?.getOpenAPISecurityScheme(scheme)
+    }
+
+    @JsonIgnore
+    override fun getBasicAuthSecurityToken(
+        schemeName: String,
+        securitySchemeConfiguration: SecuritySchemeConfiguration?
+    ): String? {
+        val tokenFromConfig = (securitySchemeConfiguration as? BasicAuthSecuritySchemeConfiguration)?.token
+        return resolveSecurityToken(tokenFromConfig, schemeName, SPECMATIC_BASIC_AUTH_TOKEN)
+    }
+
+    @JsonIgnore
+    override fun getBearerSecurityToken(
+        schemeName: String,
+        securitySchemeConfiguration: SecuritySchemeConfiguration?
+    ): String? {
+        val tokenFromConfig = (securitySchemeConfiguration as? SecuritySchemeWithOAuthToken)?.token
+        return resolveSecurityToken(tokenFromConfig, schemeName, SPECMATIC_OAUTH2_TOKEN)
+    }
+
+    @JsonIgnore
+    override fun getApiKeySecurityToken(
+        schemeName: String,
+        securitySchemeConfiguration: SecuritySchemeConfiguration?
+    ): String? {
+        val tokenFromConfig = (securitySchemeConfiguration as? APIKeySecuritySchemeConfiguration)?.value
+        return resolveSecurityToken(tokenFromConfig, schemeName)
+    }
+
+    private fun resolveSecurityToken(
+        tokenFromConfig: String?,
+        schemeName: String,
+        defaultEnvVar: String? = null
+    ): String? {
+        val configuredToken = tokenFromConfig?.takeIf { it.isNotBlank() }
+        return configuredToken
+            ?: getStringValue(schemeName)
+            ?: defaultEnvVar?.let { getStringValue(it) }
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceHost(): String? {
+        return getVirtualServiceConfiguration(this).getHost()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServicePort(): Int? {
+        return getVirtualServiceConfiguration(this).getPort()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceSpecs(): List<String>? {
+        return getVirtualServiceConfiguration(this).getSpecs()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceSpecsDirPath(): String? {
+        return getVirtualServiceConfiguration(this).getSpecsDirPath()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceLogsDirPath(): String? {
+        return getVirtualServiceConfiguration(this).getLogsDirPath()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceLogMode(): VirtualServiceConfiguration.VSLogMode? {
+        return getVirtualServiceConfiguration(this).getLogMode()
+    }
+
+    @JsonIgnore
+    override fun getVirtualServiceNonPatchableKeys(): Set<String> {
+        return getVirtualServiceConfiguration(this).getNonPatchableKeys()
+    }
+
+    @JsonIgnore
+    override fun stubContracts(relativeTo: File): List<String> {
+        return sources.flatMap { source ->
+            source.stub.orEmpty().flatMap { stub ->
+                stub.specs()
+            }.map { spec ->
+                if (source.provider == web) spec
+                else spec.canonicalPath(relativeTo)
+            }
+        }
+    }
+
+    override fun updateReportConfiguration(reportConfiguration: ReportConfiguration): SpecmaticConfigV1V2Common {
+        val reportConfigurationDetails = reportConfiguration as? ReportConfigurationDetails ?: return this
+        return this.copy(report = reportConfigurationDetails)
+    }
+
+    override fun getEnvironment(envName: String): JSONObjectValue {
+        val envConfigFromFile = environments?.get(envName) ?: return JSONObjectValue()
+
+        try {
+            return parsedJSONObject(content = ObjectMapper().writeValueAsString(envConfigFromFile))
+        } catch(e: Throwable) {
+            throw ContractException("Error loading Specmatic configuration: ${e.message}")
+        }
+    }
+
+    override fun enableResiliencyTests(): SpecmaticConfigV1V2Common {
+        val testConfig = test ?: TestConfiguration()
+        return this.copy(
+            test = testConfig.copy(
+                resiliencyTests = (testConfig.resiliencyTests ?: ResiliencyTestsConfig()).copy(
+                    enable = ResiliencyTestSuite.all,
+                ),
+            ),
+        )
+    }
+
+    override fun withTestModes(strictMode: Boolean?, lenientMode: Boolean): SpecmaticConfigV1V2Common {
+        val testConfig = test ?: TestConfiguration()
+        return this.copy(
+            test = testConfig.copy(
+                strictMode = strictMode ?: testConfig.strictMode,
+                lenientMode = lenientMode,
+            ),
+        )
+    }
+
+    override fun withTestFilter(filter: String?): SpecmaticConfigV1V2Common {
+        if (filter == null) return this
+        val testConfig = this.test ?: TestConfiguration()
+        return this.copy(test = testConfig.copy(filter = filter))
+    }
+
+    override fun withTestTimeout(timeoutInMilliseconds: Long?): SpecmaticConfigV1V2Common {
+        if (timeoutInMilliseconds == null) return this
+        val testConfig = this.test ?: TestConfiguration()
+        return this.copy(test = testConfig.copy(timeoutInMilliseconds = timeoutInMilliseconds))
+    }
+
+    override fun withStubModes(strictMode: Boolean?): SpecmaticConfigV1V2Common {
+        if (strictMode == null) return this
+        val stubConfig = this.stub ?: StubConfiguration()
+        return this.copy(stub = stubConfig.copy(strictMode = strictMode))
+    }
+
+    override fun withStubFilter(filter: String?): SpecmaticConfigV1V2Common {
+        if (filter == null) return this
+        val stubConfig = this.stub ?: StubConfiguration()
+        return this.copy(stub = stubConfig.copy(filter = filter))
+    }
+
+    override fun withGlobalMockDelay(delayInMilliseconds: Long): SpecmaticConfigV1V2Common {
+        val stubConfig = this.stub ?: StubConfiguration()
+        return this.copy(stub = stubConfig.copy(delayInMilliseconds = delayInMilliseconds))
+    }
+
+    @JsonIgnore
+    override fun testSpecPathFromConfigFor(absoluteSpecPath: String): String? {
+        val source = testSourceFromConfig(absoluteSpecPath) ?: return null
+        return source.specsUsedAsTest().firstOrNull {
+            absoluteSpecPath.contains(it)
+        }
+    }
+
+    @JsonIgnore
+    override fun stubSpecPathFromConfigFor(absoluteSpecPath: String): String? {
+        val source = stubSourceFromConfig(absoluteSpecPath) ?: return null
+        return source.specsUsedAsStub().firstOrNull {
+            absoluteSpecPath.contains(it)
+        }
+    }
+
+    @JsonIgnore
+    private fun testSourceFromConfig(absoluteSpecPath: String): Source? {
+        return sources.firstOrNull { source ->
+            source.test.orEmpty().any { test -> test.contains(absoluteSpecPath) }
+        }
+    }
+
+    @JsonIgnore
+    private fun stubSourceFromConfig(absoluteSpecPath: String): Source? {
+        return sources.firstOrNull { source ->
+            source.stub.orEmpty().any { stub -> stub.contains(absoluteSpecPath) }
+        }
+    }
+
+    @JsonIgnore
+    private fun String.canonicalPath(relativeTo: File): String {
+        return relativeTo.parentFile?.resolve(this)?.canonicalPath ?: File(this).canonicalPath
+    }
+
+    @JsonIgnore
+    override fun getLicensePath(): Path? {
+        return licensePath
+    }
+
+    @JsonIgnore
+    override fun getReportDirPath(suffix: String?): Path {
+        val reportDirPath = reportDirPath ?: defaultReportDirPath
+        return if (suffix == null) reportDirPath
+        else reportDirPath.resolve(suffix)
+    }
+
+    override fun plusExamples(exampleDirectories: List<String>): SpecmaticConfig {
+        return copy(examples = getExamples().plus(exampleDirectories))
+    }
+
+    @JsonIgnore
+    override fun getSecurityConfiguration(): SecurityConfiguration? {
+        return security
+    }
+
+    private fun File.sameAs(other: File): Boolean {
+        return try {
+            this.toPath().toRealPath() == other.toPath().toRealPath()
+        } catch (_: Exception) {
+            this.canonicalFile == other.canonicalFile
+        }
+    }
+}
+
+data class TestConfiguration(
+    val resiliencyTests: ResiliencyTestsConfig? = null,
+    val validateResponseValues: Boolean? = null,
+    val allowExtensibleSchema: Boolean? = null,
+    val timeoutInMilliseconds: Long? = null,
+    val strictMode: Boolean? = null,
+    val lenientMode: Boolean? = null,
+    val parallelism: String? = null,
+    val maxTestRequestCombinations: Int? = null,
+    val maxTestCount: Int? = null,
+    val testsDirectory: String? = null,
+    val swaggerUrl: String? = null,
+    val swaggerUIBaseURL: String? = null,
+    val actuatorUrl: String? = null,
+    val filter: String? = null,
+    val baseUrl: String? = null,
+    val filterName: String? = null,
+    val filterNotName: String? = null,
+    val overlayFilePath: String? = null,
+    val junitReportDir: String? = null,
+)
+
+enum class ResiliencyTestSuite {
+    all, positiveOnly, none
+}
+
+data class ResiliencyTestsConfig(
+    val enable: ResiliencyTestSuite? = null
+) {
+    constructor(isResiliencyTestFlagEnabled: Boolean, isOnlyPositiveFlagEnabled: Boolean) : this(
+        enable = getEnableFrom(isResiliencyTestFlagEnabled, isOnlyPositiveFlagEnabled)
+    )
+
+    companion object {
+        fun fromSystemProperties() = ResiliencyTestsConfig(
+            isResiliencyTestFlagEnabled = getBooleanValue(SPECMATIC_GENERATIVE_TESTS),
+            isOnlyPositiveFlagEnabled = getBooleanValue(ONLY_POSITIVE)
+        )
+
+        private fun getEnableFrom(
+            isResiliencyTestFlagEnabled: Boolean,
+            isOnlyPositiveFlagEnabled: Boolean
+        ): ResiliencyTestSuite? {
+            return when {
+                isResiliencyTestFlagEnabled -> ResiliencyTestSuite.all
+                isOnlyPositiveFlagEnabled -> ResiliencyTestSuite.positiveOnly
+                else -> null
+            }
+        }
+    }
+}
+
+data class Auth(
+    @param:JsonProperty("bearer-file") val bearerFile: String = "bearer.txt",
+    @param:JsonProperty("bearer-environment-variable") val bearerEnvironmentVariable: String? = null,
+    @param:JsonProperty("personal-access-token") @JsonAlias("personalAccessToken") val personalAccessToken: String? = null
+)
+
+enum class PipelineProvider { azure }
+
+data class Pipeline(
+    private val provider: PipelineProvider = PipelineProvider.azure,
+    private val organization: String = "",
+    private val project: String = "",
+    private val definitionId: Int = 0
+) {
+    fun getProvider(): PipelineProvider {
+        return provider
+    }
+
+    fun getOrganization(): String {
+        return organization
+    }
+
+    fun getProject(): String {
+        return project
+    }
+
+    fun getDefinitionId(): Int {
+        return definitionId
+    }
+}
+
+data class Environment(
+    val baseurls: Map<String, String>? = null,
+    val variables: Map<String, String>? = null
+)
+
+enum class SourceProvider { git, filesystem, web }
+
+data class Source(
+    @field:JsonAlias("type")
+    val provider: SourceProvider = filesystem,
+    val repository: String? = null,
+    val branch: String? = null,
+    @field:JsonAlias("provides")
+    @JsonDeserialize(using = ConsumesDeserializer::class)
+    val test: List<SpecExecutionConfig>? = null,
+    @field:JsonAlias("consumes")
+    @JsonDeserialize(using = ConsumesDeserializer::class)
+    val stub: List<SpecExecutionConfig>? = null,
+    val directory: String? = null,
+    val matchBranch: Boolean? = null,
+) {
+    constructor(test: List<String>? = null, stub: List<String>? = null) : this(
+        test = test?.map { SpecExecutionConfig.StringValue(it) },
+        stub = stub?.map { SpecExecutionConfig.StringValue(it) }
+    )
+
+    fun specsUsedAsStub(): List<String> {
+        return stub.orEmpty().flatMap { it.specs() }
+    }
+
+    fun specsUsedAsTest(): List<String> {
+        return test?.flatMap { it.specs() } ?: test.orEmpty().flatMap { it.specs() }
+    }
+
+    fun specToStubBaseUrlMap(defaultBaseUrl: String? = null): Map<String, String?> {
+        return stub.orEmpty().flatMap {
+            it.specToBaseUrlPairList(defaultBaseUrl) { configValue ->
+                if(configValue.specType == SpecType.OPENAPI.value) OpenAPIMockConfig.from(configValue.config).baseUrl
+                else null
+            }
+        }.toMap()
+    }
+
+    fun specToTestBaseUrlMap(defaultBaseUrl: String? = null): Map<String, String?> {
+        val baseUrlFromConfig : (SpecExecutionConfig.ConfigValue) -> String? = {
+            if(it.specType == SpecType.OPENAPI.value) OpenAPITestConfig.from(it.config).baseUrl
+            else null
+        }
+        return test?.flatMap {
+            it.specToBaseUrlPairList(defaultBaseUrl, baseUrlFromConfig)
+        }?.toMap() ?: test.orEmpty().flatMap {
+            it.specToBaseUrlPairList(defaultBaseUrl, baseUrlFromConfig)
+        }.toMap()
+    }
+
+    fun specToTestGenerativeMap(): Map<String, ResiliencyTestSuite?> {
+        return test?.flatMap {
+            when (it) {
+                is SpecExecutionConfig.StringValue -> listOf(it.value to null)
+                is SpecExecutionConfig.ObjectValue -> it.specs.map { specPath ->
+                    specPath to it.resiliencyTests?.enable
+                }
+                is SpecExecutionConfig.ConfigValue  -> it.specs.map { specPath ->
+                    if(it.specType == SpecType.OPENAPI.value) specPath to OpenAPITestConfig.from(it.config).resiliencyTests?.enable
+                    else specPath to null
+                }
+            }
+        }?.toMap() ?: emptyMap()
+    }
+
+    fun specToTestExamplesMap(): Map<String, List<String>> {
+        return test?.flatMap {
+            if(it is SpecExecutionConfig.ConfigValue && it.specType == SpecType.OPENAPI.value) {
+                return@flatMap it.specs.map { specPath ->
+                    specPath to OpenAPITestConfig.from(it.config).examples.orEmpty()
+                }
+            }
+            emptyList()
+        }?.toMap().orEmpty()
+    }
+
+    fun specToStubExamplesMap(): Map<String, List<String>> {
+        return stub?.flatMap {
+            if(it is SpecExecutionConfig.ConfigValue && it.specType == SpecType.OPENAPI.value) {
+                return@flatMap it.specs.map { specPath ->
+                    specPath to OpenAPIMockConfig.from(it.config).examples.orEmpty()
+                }
+            }
+            emptyList()
+        }?.toMap().orEmpty()
+    }
+
+
+    fun getCanonicalTestConfigs(): List<SpecExecutionConfig> {
+        if (this.test == null) return emptyList()
+        val baseDirectory = getBaseDirectory()
+        return this.test.map { config ->  config.resolveAgainst(baseDirectory) }
+    }
+
+    fun getCanonicalStubConfigs(): List<SpecExecutionConfig> {
+        if (this.stub == null) return emptyList()
+        val baseDirectory = getBaseDirectory()
+        return this.stub.map { config ->  config.resolveAgainst(baseDirectory) }
+    }
+
+    private fun getBaseDirectory(): File {
+        val currentWorkingDirectory = File(".").canonicalFile
+        return when (provider) {
+            SourceProvider.web -> currentWorkingDirectory
+            SourceProvider.filesystem -> directory?.let(::File) ?: currentWorkingDirectory
+            SourceProvider.git -> {
+                val repository = repository?.split("/")?.lastOrNull()?.removeSuffix(".git")
+                if (repository != null) {
+                    currentWorkingDirectory.resolve(".specmatic").resolve("repos").resolve(repository)
+                } else {
+                    currentWorkingDirectory
+                }
+            }
+        }
+    }
+}
+
+data class RepositoryInfo(
+    private val provider: String,
+    private val collectionName: String
+) {
+    fun getProvider(): String {
+        return provider
+    }
+
+    fun getCollectionName(): String {
+        return collectionName
+    }
+}
+
+interface ReportConfiguration {
+    fun getSuccessCriteria(): SuccessCriteria
+    fun excludedOpenAPIEndpoints(): List<String>
+
+    companion object {
+        val default = ReportConfigurationDetails(
+           types = ReportTypes()
+        )
+    }
+}
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+data class ReportConfigurationDetails(
+    val types: ReportTypes? = null
+) : ReportConfiguration {
+
+    fun validatePresenceOfExcludedEndpoints(currentVersion: SpecmaticConfigVersion): ReportConfigurationDetails {
+        if(currentVersion.isLessThanOrEqualTo(VERSION_1))
+            return this
+
+        if (types?.apiCoverage?.openAPI?.excludedEndpoints.orEmpty().isNotEmpty()) {
+            throw UnsupportedOperationException(excludedEndpointsWarning)
+        }
+        return this
+    }
+
+    fun clearPresenceOfExcludedEndpoints(): ReportConfigurationDetails {
+        return this.copy(
+            types = types?.copy(
+                apiCoverage = types.apiCoverage?.copy(
+                    openAPI = types.apiCoverage.openAPI?.copy(
+                        excludedEndpoints = emptyList()
+                    )
+                )
+            )
+        )
+    }
+
+    @JsonIgnore
+    override fun getSuccessCriteria(): SuccessCriteria {
+        return types?.apiCoverage?.openAPI?.successCriteria ?: SuccessCriteria.default
+    }
+
+    @JsonIgnore
+    override fun excludedOpenAPIEndpoints(): List<String> {
+        return types?.apiCoverage?.openAPI?.excludedEndpoints ?: emptyList()
+    }
+}
+
+data class ReportTypes(
+    @param:JsonProperty("APICoverage")
+    val apiCoverage: APICoverage? = null
+)
+
+data class APICoverage(
+    @param:JsonProperty("OpenAPI")
+    val openAPI: APICoverageConfiguration? = null
+)
+
+data class APICoverageConfiguration(
+    val successCriteria: SuccessCriteria? = null,
+    val excludedEndpoints: List<String>? = null
+)
+
+data class SuccessCriteria(
+    val minThresholdPercentage: Int? = null,
+    val maxMissedEndpointsInSpec: Int? = null,
+    val enforce: Boolean? = null
+) {
+    companion object {
+        val default = SuccessCriteria(0, 0, false)
+    }
+
+    @JsonIgnore
+    fun getMinThresholdPercentageOrDefault(): Int {
+        return minThresholdPercentage ?: 0
+    }
+
+    @JsonIgnore
+    fun getMaxMissedEndpointsInSpecOrDefault(): Int {
+        return maxMissedEndpointsInSpec ?: 0
+    }
+
+    @JsonIgnore
+    fun getEnforceOrDefault(): Boolean {
+        return enforce ?: false
+    }
+}
+
+data class SecurityConfiguration(
+    @param:JsonProperty("OpenAPI")
+    private val OpenAPI: OpenAPISecurityConfiguration?
+) {
+    fun getOpenAPISecurityScheme(scheme: String): SecuritySchemeConfiguration? {
+        return OpenAPI?.securitySchemes?.get(scheme)
+    }
+}
+
+data class OpenAPISecurityConfiguration(
+    val securitySchemes: Map<String, SecuritySchemeConfiguration> = emptyMap()
+)
+
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
+@JsonSubTypes(
+    JsonSubTypes.Type(value = OAuth2SecuritySchemeConfiguration::class, name = "oauth2"),
+    JsonSubTypes.Type(value = BasicAuthSecuritySchemeConfiguration::class, name = "basicAuth"),
+    JsonSubTypes.Type(value = BearerSecuritySchemeConfiguration::class, name = "bearer"),
+    JsonSubTypes.Type(value = APIKeySecuritySchemeConfiguration::class, name = "apiKey")
+)
+sealed class SecuritySchemeConfiguration {
+    abstract val type: String
+}
+
+interface SecuritySchemeWithOAuthToken {
+    val token: String
+}
+
+@JsonTypeName("oauth2")
+data class OAuth2SecuritySchemeConfiguration(
+    override val type: String = "oauth2",
+    override val token: String = ""
+) : SecuritySchemeConfiguration(), SecuritySchemeWithOAuthToken
+
+@JsonTypeName("basicAuth")
+data class BasicAuthSecuritySchemeConfiguration(
+    override val type: String = "basicAuth",
+    val token: String = ""
+) : SecuritySchemeConfiguration()
+
+@JsonTypeName("bearer")
+data class BearerSecuritySchemeConfiguration(
+    override val type: String = "bearer",
+    override val token: String = ""
+) : SecuritySchemeConfiguration(), SecuritySchemeWithOAuthToken
+
+@JsonTypeName("apiKey")
+data class APIKeySecuritySchemeConfiguration(
+    override val type: String = "apiKey",
+    val value: String = ""
+) : SecuritySchemeConfiguration()
+
+fun loadSpecmaticConfigOrDefault(configFileName: String? = null): SpecmaticConfig {
+    return loadSpecmaticConfigOrNull(configFileName).orDefault()
+}
+
+fun loadSpecmaticConfigOrNull(configFileName: String? = null): SpecmaticConfig? =
+    loadSpecmaticConfigOrNull(configFileName, explicitlySpecifiedByUser = false)
+
+fun loadSpecmaticConfigOrNull(
+    configFileName: String? = null,
+    explicitlySpecifiedByUser: Boolean = false
+): SpecmaticConfig? {
+    return if (configFileName == null) {
+        SpecmaticConfigV1V2Common()
+    } else {
+        try {
+            loadSpecmaticConfig(configFileName)
+        } catch (e: ContractException) {
+            val message = exceptionCauseMessage(e)
+            val configFile = File(configFileName)
+
+            if (!configFile.exists() && !explicitlySpecifiedByUser) {
+                logger.debug(message)
+            } else {
+                logger.log(message)
+            }
+
+            null
+        }
+    }
+}
+
+fun loadSpecmaticConfig(configFileName: String? = null): SpecmaticConfig {
+    val configFile = File(configFileName ?: configFilePath)
+
+    if (!configFile.exists()) {
+        throw ContractException("Could not find the Specmatic configuration at path ${configFile.canonicalPath}")
+    }
+
+    try {
+        return configFile.toSpecmaticConfig()
+    } catch(e: DatabindException) {
+        throw Exception(configErrorMessage(e))
+    } catch (e: Throwable) {
+        throw Exception("Error parsing config: ${e.message}")
+    }
+}
+
+fun loadSpecmaticConfigIfAvailableElseDefault(configFileName: String? = null): SpecmaticConfig {
+    val configFile = File(configFileName ?: configFilePath).canonicalFile
+    return loadSpecmaticConfigOrDefault(configFile.canonicalPath)
+}
+
+fun configErrorMessage(e: DatabindException): String {
+    val location = if(e.location?.lineNr != null && e.location?.columnNr != null) {
+        " (line ${e.location.lineNr}, column ${e.location.columnNr})"
+    } else {
+        ""
+    }
+
+    val errorPrefix = "Error parsing config$location"
+
+    return when (e) {
+        is InvalidNullException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            "$errorPrefix: $fieldPath must not be null, but found null."
+        }
+
+        is InvalidFormatException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            val actualValueClause = e.value?.let { " ($it is invalid)" }
+            "$errorPrefix: $fieldPath accepts ${expectedType(e)}$actualValueClause"
+        }
+
+        is IgnoredPropertyException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            "$errorPrefix: $fieldPath is not a valid property in the configuration file."
+        }
+
+        is UnrecognizedPropertyException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            val knownProperties = e.knownPropertyIds.sortedBy { it.toString() }.joinToString(", ")
+            "$errorPrefix: $fieldPath is not a valid property in the configuration file. Known properties are: $knownProperties."
+        }
+
+        is MismatchedInputException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            "$errorPrefix: $fieldPath accepts ${expectedType(e)}"
+        }
+
+        is JsonMappingException -> {
+            val path = e.path
+            val fieldPath = readablePath(path)
+            "$errorPrefix at $fieldPath: ${e.originalMessage}"
+        }
+
+        else -> {
+            "$errorPrefix: ${e.originalMessage}"
+        }
+    }
+}
+
+private fun expectedType(e: MismatchedInputException): String {
+    val targetType = e.targetType ?: return "an object"
+
+    return if (targetType.isEnum) {
+        "one of the following: ${targetType.enumConstants.orEmpty().joinToString(", ") { it.toString() }}"
+    } else if (targetType.isPrimitive) {
+        targetType.simpleName.orEmpty().lowercase()
+    } else if(targetType.genericSuperclass in listOf(java.lang.Number::class.java, java.lang.String::class.java, java.lang.Boolean::class.java)) {
+        targetType.simpleName.orEmpty().lowercase()
+    } else if (targetType.simpleName == "ArrayList") {
+        "a list"
+    } else {
+        "an object"
+    }
+}
+
+private fun readablePath(path: MutableList<JsonMappingException.Reference>) =
+    path.joinToString(".") { it.fieldName ?: "[${it.index}]" }.replace(".[", "[")
+
+fun readEnvVarOrProperty(
+    envVarName: String,
+    propertyName: String,
+): String? = System.getenv(envVarName) ?: System.getProperty(propertyName)
