+++ core/src/main/kotlin/io/specmatic/core/config/SpecmaticVersionedConfig.kt
@@ -1,7 +1,7 @@
 package io.specmatic.core.config
 
-import io.specmatic.core.SpecmaticConfig
+import io.specmatic.core.SpecmaticConfigV1V2Common
 
 interface SpecmaticVersionedConfig {
-    fun transform(): SpecmaticConfig
+    fun transform(): SpecmaticConfigV1V2Common
 }
