+++ core/src/test/kotlin/integration_tests/OpenApi31Test.kt
@@ -5,6 +5,7 @@ import io.specmatic.core.ResiliencyTestSuite
 import io.specmatic.core.ResiliencyTestsConfig
 import io.specmatic.core.Result
 import io.specmatic.core.SpecmaticConfig
+import io.specmatic.core.SpecmaticConfigV1V2Common
 import io.specmatic.core.StubConfiguration
 import io.specmatic.core.TestConfiguration
 import io.specmatic.core.examples.server.ScenarioFilter
@@ -42,7 +43,7 @@ class OpenApi31Test {
     private val openApi31File = File("src/test/resources/openapi/3_1/mixed/openapi31.yaml")
     private val openApi30File = File("src/test/resources/openapi/3_1/mixed/openapi30.yaml")
     private val backwardCompatibilityFilter = "PATH!='/multiTypeEnumSchema'"
-    private val specmaticConfig = SpecmaticConfig(
+    private val specmaticConfig = SpecmaticConfigV1V2Common(
         test = TestConfiguration(resiliencyTests = ResiliencyTestsConfig(enable = ResiliencyTestSuite.all)),
         stub = StubConfiguration(generative = true),
     )
