+++ junit5-support/src/main/kotlin/io/specmatic/test/SpecmaticJUnitSupport.kt
@@ -3,7 +3,6 @@ package io.specmatic.test
 import io.specmatic.conversions.OpenApiSpecification
 import io.specmatic.conversions.convertPathParameterStyle
 import io.specmatic.core.*
-import io.specmatic.core.SpecmaticConfig.Companion.getSecurityConfiguration
 import io.specmatic.core.filters.ScenarioMetadataFilter
 import io.specmatic.core.filters.ScenarioMetadataFilter.Companion.filterUsing
 import io.specmatic.core.log.LogMessage
@@ -349,7 +348,7 @@ open class SpecmaticJUnitSupport {
                             contractPathData.repository,
                             contractPathData.branch,
                             contractPathData.specificationPath,
-                            getSecurityConfiguration(specmaticConfig),
+                            specmaticConfig.getSecurityConfiguration(),
                             filterName,
                             filterNotName,
                             specmaticConfig = specmaticConfig,
